+++ cpp/opendnp3tests/src/AppLayerTest.h
@@ -37,7 +37,7 @@ namespace opendnp3
 class AppLayerTest
 {
 public:
-	AppLayerTest(bool aIsMaster = false, size_t aNumRetry = 0, LogLevel aLevel = LogLevel::Warning, bool aImmediate = false);
+	AppLayerTest(bool aIsMaster = false, uint32_t aNumRetry = 0, LogLevel aLevel = LogLevel::Warning, bool aImmediate = false);
 
 	void SendUp(const std::string& aBytes);
 	void SendUp(AppControlField control, FunctionCode aCode);
